LayoutTests/imported/w3c/web-platform-tests/xhr/formdata-expected.txt





FAIL Constructors should throw a type error assert_throws_js: function "() => { new FormData(null); }" did not throw
PASS empty formdata 
PASS formdata with string 
PASS formdata with named string 
PASS formdata from form 
FAIL Newly created FormData contains entries added to "formData" IDL attribute of FormDataEvent. assert_equals: expected (string) "vh1" but got (object) null
FAIL |new FormData()| in formdata event handler should throw assert_equals: expected 1 but got 0
